ingestion/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pymupdf==1.24.10
google-cloud-bigquery==3.25.0
google-cloud-storage==2.18.2
google-cloud-documentai==2.31.0
google-cloud-aiplatform==1.71.1
tiktoken==0.7.0
